Move `Publish` to low-level API
