+++ identity-iota/src/tangle/mod.rs
@@ -18,6 +18,7 @@ pub use self::message_ext::TryFromMessage;
 pub use self::message_index::MessageIndex;
 pub use self::network::Network;
 pub use self::network::NetworkName;
+pub use self::publish::Publish;
 pub use self::receipt::Receipt;
 pub use self::traits::TangleRef;
 pub use self::traits::TangleResolve;
@@ -28,5 +29,6 @@ mod client_map;
 mod message_ext;
 mod message_index;
 mod network;
+mod publish;
 mod receipt;
 mod traits;
